fix: consider default style when computing effective style properties
The StyleSheet.getCellStyle didn't keep the properties of the default style when `baseStyleNames` was set in the
`cellStyle` parameter.

The JSDoc was incorrect (it came from mxGraph) about how the style is computed, in particular about the default style.
It probably leads to the erroneous implementation during migration. It is now fixed and clearly describe the rules
followed to merge style properties.

The 'Stylesheet' has been updated to correctly use the maxGraph API. It also includes more examples involving
`baseStyleNames` to show the 'properties merge' in action.

Introduce `jest` to test the fix and the whole implementation of the changed method.
Types check support is provided by `ts-jest`. As maxGraph uses a lot of browser objects, also setup
`jest-jsdom-environment`.