fix: CellOverlay constructor use all parameters to set properties
The `align` and `verticalAlign` values passed to the constructor weren't set, so the overlay position was always set to
the default one.

The `align` and `verticalAlign` properties are now using the AlignValue and VAlignValue types respectively instead of
defining inline types. This improves the consistent in the whole code.

To demonstrate the fix, the Overlays story now set 'align' and 'verticalAlign' randomly.