fix: improve style configuration extensibility
The type of the `endArrow`, `startArrow` and `shape` properties in `CellStateStyle` were too strict. They prevented to
use custom registered shapes in the style configuration without using a cast.

Markers of the arrow and custom shapes can be registered by using methods in `MarkerShape` and `CellRenderer` respectively.
The key to identify the shapes is a string so the types of the properties in `CellStateStyle` now allow `string` as well.
These offer both the type guidance for standard maxGraph registered shapes (via a dedicated type) and the extensibility
for custom shapes.