test: execute tests serially to speed up test execution
By default, `jest` runs tests in parallel. However, with the current configuration of `ts-jest`, all workers run `tsc`
in parallel, which is very slow and consumes a lot of resources.
This is what takes the longest.
By running `jest` serially, `tsc` runs before starting the first test, then all subsequent tests run faster.
This will probably be improved by customizing the configuration of `jest` or with future versions of `ts-jest`.