refactor!: move codecs to dedicated files and reduce side effects
Previously, codecs classes were defined and registered in the same file as the class they target.
Such registration prevented an efficient tree-shaking: as far as the target class was imported (for example StyleSheet),
the related codec was imported as well. As a consequence, a large amount of codecs code (and their dependant code) was
present in applications integrating maxGraph even if they didn't use the codec feature. As there are currently a lot of
issue with the codec feature making it almost unusable, that was a large waste.

Codecs are no longer registered by default which reduce side effects and coupling.
This also improves the tree-shaking a lot.
For example, in the `ts-example` application included in this repository that doesn't use the "codec" feature, the size
of the minified maxGraph chunk decreases from `568.58 kB` to `467.90 kB`, corresponding to a `17.7%` size decrease!

BREAKING CHANGE: Codecs supplied by `maxGraph` are no longer registered by default. They MUST be registered before
performing an `encode` or `decode`. You can use one of the following functions to register codecs:
  - `registerAllCodecs`
  - `registerCoreCodecs`
  - `registerEditorCodecs`