fix: import/decode the XML Graph model when the code is minified
Create a dedicated example with webpack to reproduce the problem.
Names used to retrieve the codecs are set directly in the codecs at initialization to not depend on the actual
constructor names.

This changes also partially fixe the export/encode. The Geometry is not correctly encoded.
More work is needed to make this work, probably by modifying the way Codecs are registered and retrieved.