fix: use correct perimeter style property in stories
Some stories used a perimiter instead of perimeter. This is now fixed.
When possible, do not set the RectanglePerimeter in the style configuration as it is already set
in the default vertex style.
Migrate some stories to TypeScript when possible (migrating Swimlanes involved too many unrelated
changes, so it will be done later).

The following stories already had issues unrelated to the perimeter property and are still not
working. This is why they weren't migrated to TypeScript as part of the fix.
  - FileIO
  - Monitor
  - Scrollbars