feat: add a story to demonstrate all Perimeter features
Demonstrate the effect of all available perimeters on various shapes and what happens when no
perimeter is configured.

Migrate existing Perimeter history to TypeScript.
It also stops modifying the GraphView prototype, extending the class instead. This avoids side
effects in other stories (as they run in a Single-Page application).
This story is still not fully working: the perimeter doesn't apply to the label until the shape is
rendered another time (by moving it for example). This worked in the mxGraph example and should be
later investigate.