refactor: migrate the DynamicToolbar story to TypeScript
Fix the style to ensure that a dropped element uses the correct shape. Previously, all dropped
elements used the Rectangle shape instead of the shape chosen in the toolbar
Remove an additional listener previously used for IE in mxGraph, as IE is not supported in maxGraph.
Comment on the use of the KeyHandler, which is currently not working.