fix: make usage of Codecs work in stories
Previously, some stories still encode/decode by calling directly the Codec code. Default Codecs
were not registered so errors occurred.
Most of the stories updated as part of the fix have been migrated to TypesScript to also fix other
minor issues.

Clipboard (still in JavaScript)
The Clipboard actions still don't work and have not been touched as this is out of the scope of the
changes proposed here.
The code that uses the Codecs to generate the xml string used for copy/paste (the string
representation of the cells to copy) is now working (the Codec are registered).

FileIO
  - load of txt and xml file works
  - the user can choose the source of the model with a Storybook control
  - display images
  - display the "arrange" button

JSonData
  - display the content
  - display the popup

UserObject
  - load images (for popup)
  - fix style and form div
  - improve the layout
  - fix error when validating new value in the form (wrong import)

DISCLAIMER about the maxGraph "popup"
The CSS class are still not correctly set in the maxGraph code, so popup still don't display
correctly.
The popup are not cleaned when changing from a story to another, as they are currently not created
in the DOM at the right place. They are created directly in whole document, so not at the place
where the story is located in the DOM (not in the right div).