fix: correctly set the flag value in `setStyleFlag`
Previously, when the style value was undefined for the given key, the function set the style value
to 0 when the value parameter was true instead of setting the flag value.
This problem had been introduced during the migration from `mxGraph`.

The function is now fully covered by automatic tests to validate that the entire implementation
behaves correctly.

In addition:
  - improve the JSDoc of the `setCellStyleFlag` and `setCellStyleFlags` functions
  - fix the signature of the `setCellStyleFlags` function. The `value` parameter is optional.