fix: correctly compute the actual style when setting undefined/non values
Previously, when the value of a property was defined as `undefined`, it was also defined as `undefined`
in the calculated style. This was a change from the mxGraph implementation and was not
consistent with the style without the same property.
This has now been corrected: when a value is defined as undefined, the value calculated is that of
the default style or a "base" style.
To match mxGraph's behavior, the same logic applies when the value is set to "none".