fix: correctly detect parent in SwimlaneMixin.dropTarget
The target was not detected as a valid target so, the targeted swimlane was never highlighted while
panning the cell over the potential target.

The problem was visible in the Swimlanes and Folding stories, the targeted swimlanes are now
correctly highlighted.

The problem was introduced during the mxGraph migration, the parent variable scope change during the
migration and the parent variable used to validate if the target was valid didn't refer to a Cell in
maxGraph but to Window.parent.