fix: prevent error when not passing "mapping" to Graph.cloneCell
Previously, this generated an error in `ConnectionHandler` because the default mapping value was
invalid (null is not allowed) and the error opened a MaxLog without error message.
ConnectionHandler has been updated to better track such an error.

The `ContexIcons` story exhibited the problem. It is now fully working.
It has also been migrated to TypeScript to ease future maintenance.

Also update the `ConnectionHandler.start` method signature as part of the migration of the story to
TypeScript.
The `edgeState` parameter can now be undefined, a default value was already managed in code.