fix: set correct type for XmlCanvas2D.root
The type of the `root` property had been incorrectly set to `SVGElement` during the Typescript
migration (see commit 413796ad).
This prevented to use it in various scenario, like the example described in the `ExportImage` JSDoc.
This example was working in mxGraph and typed-mxgraph declares the property as a `Element`, so it is
now declares the property as a `Element`

A test has been added to show that the ExportImage example using XmlCanvas2D with a root using an
Element type works.