feat: declare the package as side effects free
`maxGraph` code no longer contains side effects, so it can be declared as such in the `package.json` file.
This allows bundlers to perform more efficient tree-shaking, reducing maxGraph's footprint in applications.

Declare CSS files in side effects
This ensures that when using the webpack css-loader plugin and importing a CSS file from `maxGraph`,
this CSS file is not dropped during tree-shaking.