refactor: remove unused methods in shapes
Remove `Shape.reconfigure`
This method was also unused in `mxGraph`.

Remove `ImageShape.createHtml`
It was previously called by `Shape.create` in mxGraph when the dom element creation depends on the
dialect value.
When maxGraph removed the support for former Vml and Html canvas, `Shape.create` was updated to
inline the dom node creation for SVG only.
The `ImageShape.createHtml` should have been removed at that time.