feat: provide the list of default plugins
This is useful for enriching the list of plugins passed to the `Graph` constructor.

Demonstrate the usage of the new functions in examples and in TypeScript stories.
Improve the JSDoc of `RubberbandHandler` to better explain its usage as a Plugin.
Introduce the Plugin category to better group the features relating to the feature in the HTML doc.