docs: explain Perimeter in the documentation site
This is an introduction to the concept of “perimeter” and a quick explanation of how to use perimeters.
Previously, this information was contained in the `Perimeter` JSDoc. This served as the basis for the
new documentation. This has now been removed, in order to maintain a single source of information and
facilitate maintenance.