fix: correctly call methods in TemporaryCellStates and PrintPreview
In the `PageBreaks` story, the print page failed to display the graph.
This issue was mainly caused by maxGraph core calling functions on incorrect instances, especially when switching from regular functions to arrow functions (changing the value of `this`).

This problem originated during the migration from mxGraph.

Additionally, in `TemporaryCellStates`, the correct method `oldDoRedrawShape` is now called instead of `oldValidateCellState`.
The addition of types helped in identifying this issue.

Finally, `PrintPreview` now logs errors to the global logger, allowing easier troubleshooting in the future.
The previous error was hidden, and enabling logs made it possible to trace the source of the problem.