fix: allow to insert vertex without setting position and size
This was allowed in mxGraph, and it is useful in scenario involving layouts (which is in charge of
computing the position).
The size of a vertex is sometime computed after the insertion, like in the OrgChart story.

Update the signature of both the createVertex and insertVertex methods, and improve the related JSDOC.
When the position is not set, the default values of the Geometry instance are used.

Refactor tests: introduce helpers functions to remove duplication and make the expectations easier
to read.