chore(deps-dev): bump docusaurus from 3.5.2 to 3.6.0
Enable the new faster build provided by rspack.
The build is now 2 or 3 times faster than when using docusaurus 3.5.2 version with webpack.
Do not use the rsdoctor plugin for now. There is no need for a deeper analysis, as the build is
faster than in the past.

Also bump react and other dependencies.