feat: allow usage of functions in `CellStateStyle.edgeStyle`
This was working but the type definition didn't allow it.
Some stories (mainly written in JavaScript) were already configured to use it. Some TypeScript
stories are now using it as well.

The actual support for the usage of a function was already implemented in `GraphView.getEdgeStyle`.

Also improve the JSDoc of EdgeStyle, in particular, code examples.