fix: change definition of isScrollEvent
Current definition of isScrollEvent assumes that the event "evt" passed to function would be MouseEvent.
If the evt was triggered from Touch Device(in my case iPad Air)  it would pass TouchEvent which doesn't directly have access to clientX and clientY which was causing the function to throw error "Invalid X supplied" when creating a new Point.