fix: allow the edit inside the editing story example (#544)
Properly editable is not being set by default for cells, hence the editing was never allowed to start.
Major issue that was also present inside the mxGraph old example was that the value doesn't actually get updated. This commit fixes that as well.

Furthermore, the container for the textarea editing had position relative instead of absolute which caused snappy behaviour while editing.