feat: ease configuration of handles in `EdgeHandler`
In `mxGraph`, such configuration was possible by updating some mxConstants values, but in `maxGraph`,
constants are immutable.
It is now possible to use `EdgeHandlerConfig` to globally configure various settings like the stroke
color.
Previously, configuring such settings required to extend `EdgeHandler` and its subclasses.

The `HandleConfig` object has been updated to also let configure the label handles, for example,
their size and fill color. The new properties are now used in `EdgeHandler` and `VertexHandler`.

Lastly, it is possible to reset the properties in `EdgeHandlerConfig` to its default values by using
a dedicated reset function.

A new story has been added to demonstrate the effects of customizing “handle” defaults and comparing
them with the original defaults.