docs: highlight the getting started page
Previously, the start page was located in the “manual” folder and was therefore not very visible.
This page is the first point of entry for newcomers and therefore needs to be more visible.
It has now been moved to 2nd position in the menu to give it greater prominence.

The introduction page has also been updated to link to the documentation rather than to the repository README.