chore: measure impact of a configurable i18n abstraction over Transactions on tree-shaking [POC]
Don't depend on `Translations` directly in the code.
Introduce an "I18n" abstraction as we did with Logger for MaxLog.

A configurable instance of the "I18nProvider" is set on `GlobalConfig` which is then use everywhere in the code.
The configuration related to Translations is no longer in Client but in a new dedicated `TranslationsConfig` global object.