docs: apply improvements to the navbar
Use the GitHub logo instead of a label for the link to the GitHub repository to make the right-hand side more attractive (logos only).
Hide the navigation bar when scrolling to free up screen space for reading.

Also improved the footer: “Built with Docusaurus” is now a link to the docusaurus website, to advertise this tool.