refactor!: extract Orthogonal and Manhattan connectors (#684)
The connectors are now stored in dedicated files to make the code simpler to read.
The remaining utility properties and methods in EdgeStyle were only used by the Orthogonal connector. So, they are now internal.
This somewhat reduces the size of maxGraph when it is bundled and minified in applications.

BREAKING CHANGES: properties and utility methods previously exposed by `EdgeStyle` are now only internal.
The `getRoutePattern` method has been completely removed as it was not being used. Nor was it used in the entire mxGraph and draw.io codebase.