refactor: reset more global configurations in stories
This will limit the side effects of a story's modification of a global configuration on other stories.

Do not reset the style registries (for shapes, perimeters, ...) at this time. Once they have been filled when initializing
a Graph instance, they are not filled again when initializing a new Graph instance.