refactor: use external CSS file MenuStyle and ShowRegion stories (#698)
Previously, the CSS was included programmatically in the document, which is not the best way to add CSS in the storybook context.
In this case, the CSS was added each time the story was rendered, so the CSS definition was added multiple times

The stories now use external CSS and use imports to include them.
This removes the duplicated includes in the DOM and makes the thing easier to understand by using standard practices.

Notice that the Scrollbars story is still using the former way of including custom CSS.
It is not functional and requires additional work. So, it will be updated later.