feat: allow to use custom shape for overlays (#696)
`CellRenderer` now provides 2 protected methods for deciding how to create the overlay shape and for configuring it (usually to configure the DOM node linked to the shape).

The overlay story has been modified to demonstrate how to use the new customization capabilities.
It is now also written in TypeScript for ease of maintenance.

Also improve JSDoc of `CellMarker` and `CellTracker`.