fix: consider `ManhattanConnector` as orthogonal
Currently, the implementation of a connector must be explicitly referenced in the code to decide whether it is considered orthogonal.
This is not configurable.
So, when a new connector is added, its reference must be added. This was not done when `ManhattanConnector` was introduced.

ManhattanConnector is now referenced in `Graph.isOrthogonal` and `Graph.createEdgeHandler`.
tests have been added to validate the correction and a refactoring of the implementation to simplify it.