feat: CellStateStyle port constraints accept multiple DirectionValue (#721)
Previously, the properties only accept a single DIRECTION enum value. It wasn't possible to pass a string related to the DirectionValue type, it generated a TypeScript error.
In addition, it is now possible to pass several values, as this was possible in mxGraph. This is what the Orthogonal story requires for example.

For backward compatibility, the implementation still supports the former mxGraph configuration where the directions were passed as a concatenated string. 
This ensures that the mxGraph XML models can be correctly loaded.

About 20 tests have been added to validate the existing implementation.