feat: let "fit center" with the new `FitPlugin` (#733)
Introduce `FitPlugin` which is intended to contain all fit methods in the future.

The “fit center” implementation is inspired by the old example provided in the `Graph.fit` method.
It has been updated to improve margin management and to limit rounding impacts, particularly when the method is called several times in succession.

The ts-example and the story now use the new method to demonstrate its use and operation in different contexts.