feat!: limit usage of the `eval` function (#736)
`StylesheetCodec` no longer call the function by default.
The eslint configuration now includes a new rule that prevents new calls to the function.

Introduce an internal function for indirection of the `eval` function.
This change generates fewer warnings when using bundlers such as Rollup.
Be aware that this is a temporary workaround, not a final solution.

BREAKING CHANGES: `StylesheetCodec.allowEval` is now set to `false` by default to prevent unwanted use of the eval function, as it carries a possible security risk.