feat!: introduce a way to configure the I18n provider
Previously, the i18n built-in mechanism managed by the `Translations` class was always in use.
It is now possible to plug custom i18n solutions in addition to the built-in implementation.

By default, maxGraph now use an implementation that doesn't provide any translation. This reduces the size of the
applications that don't want to use the built-in implementation.
The bundle size of examples provided in the repository has been reduced by the following values:
  - js-example: 8kB
  - ts-example: 5kB

BREAKING CHANGES: The built-in `Translations` class is no longer used by default. To use it, call `GlobalConfig.i18n = new TranslationsAsI18n()`.