chore(types): enforce usage of explicit override keyword in examples (#739)
A lot of refactoring involving methods move are planned in the near future, mainly to improve the tree shaking.
Enabling this check will detect errors early while doing such refactoring.

Do not enable it for now in the code of the lib, as we want maxGraph to support TS 3.8 and this feature requires TS 4.3.

In addition, decrease the chunk size warning limit in the "vite.config.js" file in examples (relates to a previous commit).