chore(deps-dev): bump typescript-eslint from 7.18.0 to 8.29.0 (#744)
For now, disable the "no-unsafe-function-type" rule. This requires a lot of changes which will probably will introduce breaking changes.
This rule will be activated at a later date.

Restore the automatic Dependabot update as the lint fixes related to v8 have been integrated, so it will be possible to follow recent updates.