fix(story): ensure that the Model Codecs are always registered (#749)
Manually cleaning the codecs introduced side effects, and the Codecs were unavailable once a story was visited.
For example, display a story using the Model Codecs like Monitor story, then going to another story and displaying the Monitor story again. In the latter, the model wasn't fill.

This was because `ModelSerializer` registers the Codecs but only once.
If we unregistered the Codecs manually as we do for stories, the `ModelSerializer` is currently unable to know it has to register the Codecs again (by forcing the registration).
The workaround is to force registration of Model Codecs before loading every story.