feat: simplify types used to register shapes (#768)
Previously, there were TSC errors when registering some builtin or custom shape because their constructor didn't match the signature of the `Shape` constructor.
The type of various `CellRenderer` properties and methods has been updated to fix this, so it is no longer necessary to add some 'ts-ignore' to silently ignore these false errors.

In addition:
  - Various improvements in the JSDoc of `CellRenderer`
  - `CellRenderer.getShape` allows `undefined` parameter
  - Move `ActorShape` to the directory storing vertex shapes for consistency