feat: let register and unregister default style elements (#757)
Previously, it was only possible to register the default builtins internally.
"Edge Marker" factory functions, used to display start and end arrows (and other shapes), are now public.
This makes it easier for applications to customize the style they need to configure.

Introduce "unregister" functions to clean the maxGraph registries:
  - They reset the loading states of the "default" elements to ensure they can be loaded again.
  - Codec unregister is also available.
  - All new unregister functions are used within the story to reduce edge effects between stories.