feat: add CommonJS support to npm package (#826)
The new commonjs resources are built with a new tsc command and now referenced in the package.json file.

Add new examples using CommonJS with Node.js to demonstrate the new support:
  - TypeScript example: show the integration with jest and ts-jest
  - JavaScript example: use `maxGraph` in a headless environment with Node.js (and jsdom).

This example exports Graph Model as XML, and exports the rendered Graph to SVG.

In addition, add missing git extension in the repository attribute of the package.json file.