refactor!: remove `Dictionary` and use ECMAScript `Map` instead
The `Dictionary` class had been introduced by mxGraph for browser without `Map` object support.
The `Map` class is available in the ECMAScript version supported by maxGraph, so we can remove it.

In the examples in this repository, the package size decreases by 0.7kB.

BREAKING CHANGES: The `Dictionary` class has been removed. Use the `Map` class instead, which is a standard JavaScript feature.