refactor!: remove `Dictionary` and use `Map` instead (#857)
The `Dictionary` class had been introduced by mxGraph for browser without `Map` object support.
The `Map` class is available in the ECMAScript version supported by maxGraph, so we can remove it.

In the examples in this repository, the package size decreases by 0.7kB.

BREAKING CHANGES: The `Dictionary` class has been removed. The maxGraph API use the `Map` class instead, which is a standard JavaScript feature.
If your code depends on the `Dictionary` class, you can use the `Map` class instead.