refactor: remove reference to legacy insert cells methods (#876)
These legacy methods have a large number of parameters, are hard to read and hard to evolve.

So, to prepare the deprecation of these methods, use the new methods that take a single object parameter in code, examples and documentation.
The storybook stories will be updated later.

Also simplify some calls by not always setting the parent when it was set to the default parent (this is the default
value, so no need to pass it in this case), nor setting an empty string value (same result as not setting it).

In addition, apply various JSDoc improvements.