fix: consistently compute the arcSize for edges and vertices (#884)
The arcSize of edges missed a division by 2, which was inconsistent with the past mxGraph behavior and the computation done for vertices.
This had been introduced by mistake in 0453da274fbbeba4b7a457daec1cd9ad6db4abe0.
The result of `getSwimlaneArcSize` was only correct when fallback to the default value. Otherwise, it missed the division by 2 as well.

In addition, refactor:
- remove duplication in various shapes (introduce Shape.getBaseArcSize)
- simplify implementations
- add tests
- improve JSDOC for CellStyle.arcSize

BREAKING CHANGES: The `arcSize` of rounded shapes was not always correctly computed in the past.
The computation is now consistent everywhere in the code and matches the mxGraph behavior.
To have the same rendering as before for edges, you must multiply the `arcSize` by `2` in your styles.