feat: accept more nullish parameter in various methods
This prepares the migration of the "Wires" story to TypeScript.

Add a new factory method in `CellMarker` to override the `CellHighlight` instance
Export missing `ConnectionHandlerCellMarker` class

The signature of some methods of the following classes now accepts null and/or undefined because their implementation
already supported it:
  - CellMixin
  - ConnectionHandler
  - EdgeMixin
  - Geometry.setTerminalPoint
  - GraphView.updateFixedTerminalPoint
  - VertexMixin

Improve JSDoc in various locations.