refactor: fix most problems in the Wires story (#893)
The labels are now at the right position.
Restore the original background specific to this story (dot instead of grid)
The terminal points of the R1 shape displays when hovering a edge in construction.
Let switch to a dark theme by using the Storybook controls.

Add a documentation page taking the technical explanations from the mxGraph example and that was previously hidden.
Also display a description at the top of the Graph to help understand what the story is doing.

Also migrate it to TypeScript. This facilitates maintenance and enables errors to be detected earlier.

Some refactoring
  - no longer override PanningHandler.isPopupTrigger as it does not exist (same in mxGraph). This is a method of InternalEvent
  - add nullish checks and simplify conditions